requirements.txt
amqp==2.3.2
billiard==3.5.0.4
Django==2.0.6
django-redis==4.9.0
dukpy==0.2.0
eventlet==0.23.0
future==0.16.0
greenlet==0.4.13
javascripthon==0.10
Jinja2==2.10
kombu==4.2.1
lml==0.0.2
macropy3==1.1.0b2
Markdown==3.0.1
MarkupSafe==1.0
mysqlclient==1.3.12
Pillow==5.1.0
PyJWT==1.6.4
pytz==2018.4
redis==2.10.6
vine==1.1.4
